+++ traffic_ops/traffic_ops_golang/monitoring/monitoring_test.go
@@ -151,7 +151,8 @@ func TestGetCachegroups(t *testing.T) {
 
 	mock.ExpectQuery("SELECT").WithArgs(cdn).WillReturnRows(rows)
 
-	dbCtx, _ := context.WithTimeout(context.TODO(), time.Duration(10)*time.Second)
+	dbCtx, f := context.WithTimeout(context.TODO(), time.Duration(10)*time.Second)
+	defer f()
 	tx, err := db.BeginTx(dbCtx, nil)
 	if err != nil {
 		t.Fatalf("creating transaction: %v", err)
@@ -325,7 +326,8 @@ func TestGetProfiles(t *testing.T) {
 
 	mock.ExpectQuery("SELECT").WithArgs(pq.Array(profileNames), CacheMonitorConfigFile).WillReturnRows(rows)
 
-	dbCtx, _ := context.WithTimeout(context.TODO(), time.Duration(10)*time.Second)
+	dbCtx, f := context.WithTimeout(context.TODO(), time.Duration(10)*time.Second)
+	defer f()
 	tx, err := db.BeginTx(dbCtx, nil)
 	if err != nil {
 		t.Fatalf("creating transaction: %v", err)
@@ -373,11 +375,6 @@ func TestGetDeliveryServices(t *testing.T) {
 	db := sqlx.NewDb(mockDB, "sqlmock")
 	defer db.Close()
 
-	router := Router{
-		Type:    RouterType,
-		Profile: "routerProfile",
-	}
-
 	deliveryservice := DeliveryService{
 		XMLID:              "myDsid",
 		TotalTPSThreshold:  42.42,
@@ -386,7 +383,6 @@ func TestGetDeliveryServices(t *testing.T) {
 	}
 
 	deliveryservices := []DeliveryService{deliveryservice}
-	routers := []Router{router}
 
 	mock.ExpectBegin()
 	rows := sqlmock.NewRows([]string{"xml_id", "global_max_tps", "global_max_mbps"})
@@ -394,23 +390,22 @@ func TestGetDeliveryServices(t *testing.T) {
 		rows = rows.AddRow(deliveryservice.XMLID, deliveryservice.TotalTPSThreshold, deliveryservice.TotalKBPSThreshold/KilobitsPerMegabit)
 	}
 
-	profileNames := []string{router.Profile}
+	mock.ExpectQuery("SELECT").WillReturnRows(rows)
 
-	mock.ExpectQuery("SELECT").WithArgs(pq.Array(profileNames)).WillReturnRows(rows)
-
-	dbCtx, _ := context.WithTimeout(context.TODO(), time.Duration(10)*time.Second)
+	dbCtx, f := context.WithTimeout(context.TODO(), time.Duration(10)*time.Second)
+	defer f()
 	tx, err := db.BeginTx(dbCtx, nil)
 	if err != nil {
 		t.Fatalf("creating transaction: %v", err)
 	}
 
-	sqlDeliveryservices, err := getDeliveryServices(tx, routers)
+	sqlDeliveryservices, err := getDeliveryServices(tx)
 	if err != nil {
-		t.Errorf("getProfiles expected: nil error, actual: %v", err)
+		t.Errorf("getDeliveryServices expected: nil error, actual: %v", err)
 	}
 
 	if len(deliveryservices) != len(sqlDeliveryservices) {
-		t.Errorf("getProfiles expected: %+v actual: %+v", deliveryservices, sqlDeliveryservices)
+		t.Fatalf("getDeliveryServices expected: %+v actual: %+v", deliveryservices, sqlDeliveryservices)
 	}
 
 	for i, sqlDeliveryservice := range sqlDeliveryservices {
@@ -448,7 +443,8 @@ func TestGetConfig(t *testing.T) {
 
 	mock.ExpectQuery("SELECT").WillReturnRows(rows)
 
-	dbCtx, _ := context.WithTimeout(context.TODO(), time.Duration(10)*time.Second)
+	dbCtx, f := context.WithTimeout(context.TODO(), time.Duration(10)*time.Second)
+	defer f()
 	tx, err := db.BeginTx(dbCtx, nil)
 	if err != nil {
 		t.Fatalf("creating transaction: %v", err)
@@ -610,10 +606,6 @@ func TestGetMonitoringJSON(t *testing.T) {
 		//
 		// getDeliveryServices
 		//
-		router := Router{
-			Type:    RouterType,
-			Profile: "routerProfile",
-		}
 
 		deliveryservice := DeliveryService{
 			XMLID:              "myDsid",
@@ -630,9 +622,7 @@ func TestGetMonitoringJSON(t *testing.T) {
 			rows = rows.AddRow(deliveryservice.XMLID, deliveryservice.TotalTPSThreshold, deliveryservice.TotalKBPSThreshold/KilobitsPerMegabit)
 		}
 
-		profileNames := []string{router.Profile}
-
-		mock.ExpectQuery("SELECT").WithArgs(pq.Array(profileNames)).WillReturnRows(rows)
+		mock.ExpectQuery("SELECT").WillReturnRows(rows)
 		resp.Response.DeliveryServices = deliveryservices
 	}
 	{
@@ -653,7 +643,8 @@ func TestGetMonitoringJSON(t *testing.T) {
 		resp.Response.Config = config
 	}
 
-	dbCtx, _ := context.WithTimeout(context.TODO(), time.Duration(10)*time.Second)
+	dbCtx, f := context.WithTimeout(context.TODO(), time.Duration(10)*time.Second)
+	defer f()
 	tx, err := db.BeginTx(dbCtx, nil)
 	if err != nil {
 		t.Fatalf("creating transaction: %v", err)
@@ -661,7 +652,7 @@ func TestGetMonitoringJSON(t *testing.T) {
 
 	sqlResp, err := GetMonitoringJSON(tx, cdn)
 	if err != nil {
-		t.Errorf("GetMonitoringJSON expected: nil error, actual: %v", err)
+		t.Fatalf("GetMonitoringJSON expected: nil error, actual: %v", err)
 	}
 
 	resp.Response.TrafficServers = sortCaches(resp.Response.TrafficServers)